epharm/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.2.0
Pillow-11.0.0
django-jazzmin
